parkingmanager/requirements.txt
django-createsuperuserwithpassword
django-betterforms
django==2.2.5
psycopg2
Pillow==2.6.1
pytesseract==0.3.3